separate docker archive image into its own file
